feat(core): support merge transformers
